WIP for dryrun and dirty flags

+++ gorelease.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"flag"
 	"fmt"
 	"github.com/blang/semver/v4"
 	"github.com/go-git/go-git/v5"
@@ -8,6 +9,7 @@ import (
 	"github.com/go-git/go-git/v5/plumbing"
 	"github.com/go-git/go-git/v5/plumbing/object"
 	"github.com/go-git/go-git/v5/plumbing/transport/ssh"
+	"github.com/nwillc/gorelease/version"
 	"io/ioutil"
 	"log"
 	"os"
@@ -23,32 +25,60 @@ const (
 	gitUser        = "git"
 )
 
+var dryRun *bool
+var dirty *bool
+var vers *bool
+
+func init() {
+	dryRun = flag.Bool("dryrun", false, "perform a dry run")
+	dirty = flag.Bool("dirty", false, "allow dirty repo")
+	vers = flag.Bool("version", false, "display version")
+}
+
 func main() {
+	flag.Parse()
+	if *vers {
+		fmt.Printf("version %s\n", version.Version)
+		return
+	}
+	if *dryRun {
+		log.Println("Performing dry run.")
+	}
 	repo := getRepository("")
 	w, err := repo.Worktree()
-	checkIfError(err)
+	checkIfError("repository access", err)
 	status, err := w.Status()
-	checkIfError(err)
+	checkIfError("repository status", err)
 
 	/*
 	 * Check that we are ready for release.
 	 */
 	if len(status) != 1 {
-		panic(fmt.Errorf("incorrrect file commit status, %d files, expecting only %s", len(status), dotVersionFile))
+		msg := fmt.Sprintf("incorrrect file commit status, %d files, expecting only %s", len(status), dotVersionFile)
+		if *dirty {
+			log.Println(msg)
+		} else {
+			panic(fmt.Errorf(msg))
+		}
 	}
 
 	vs := status.File(dotVersionFile)
 	if vs.Staging == '?' && vs.Worktree == '?' {
-		panic(fmt.Errorf("%s should be only uncommitted file", dotVersionFile))
+		msg := fmt.Sprintf("%s should be only uncommitted file", dotVersionFile)
+		if *dirty {
+			log.Println(msg)
+		} else {
+			panic(fmt.Errorf(msg))
+		}
 	}
 	/*
 	 * Get new version.
 	 */
 	content, err := ioutil.ReadFile(dotVersionFile)
-	checkIfError(err)
+	checkIfError("reading .version", err)
 	versionStr := strings.Replace(string(content), "\n", "", -1)
 	v, err := semver.Make(versionStr)
-	checkIfError(err)
+	checkIfError("parsing version", err)
 	tag := "v" + v.String()
 
 	/*
@@ -60,10 +90,10 @@ func main() {
 	* Git add the .version and version files.
 	 */
 	_, err = w.Add(output)
-	checkIfError(err)
+	checkIfError("adding version.go", err)
 
 	_, err = w.Add(dotVersionFile)
-	checkIfError(err)
+	checkIfError("adding .version", err)
 
 	/*
 	* Git commit the files.
@@ -71,13 +101,13 @@ func main() {
 	_, err = w.Commit("Generated for "+tag, &git.CommitOptions{
 		Author: newSignature(),
 	})
-	checkIfError(err)
+	checkIfError("committing files", err)
 
 	/*
 	* Git create new tag.
 	 */
 	ok, err := setTag(repo, tag)
-	checkIfError(err)
+	checkIfError("setting tag", err)
 
 	if !ok {
 		panic(fmt.Errorf("unable to set tag %s", tag))
@@ -115,7 +145,7 @@ func newPushOptions(refSpecs []config.RefSpec, keys *ssh.PublicKeys) *git.PushOp
 
 func publicKeys() (*ssh.PublicKeys, error) {
 	path, err := os.UserHomeDir()
-	checkIfError(err)
+	checkIfError("finding home directory", err)
 	path += "/.ssh/id_rsa"
 
 	publicKey, err := ssh.NewPublicKeysFromFile(gitUser, path, "")
@@ -127,7 +157,7 @@ func publicKeys() (*ssh.PublicKeys, error) {
 
 func newSignature() *object.Signature {
 	userInfo, err := user.Current()
-	checkIfError(err)
+	checkIfError("getting current user", err)
 	sig := object.Signature{
 		Name: userInfo.Name,
 		When: time.Now(),
@@ -137,18 +167,23 @@ func newSignature() *object.Signature {
 
 func createVersionGo(fileName string, tag string) {
 	contents, err := ioutil.ReadFile(licenseFile)
-	checkIfError(err)
+	checkIfError("reading licence", err)
 	licenseStr := strings.Replace(string(contents), "\n", "\n *", -1)
 
 	versionGo := strings.Replace(versionTemplateStr, "$LICENSE$", licenseStr, 1)
 	versionGo = strings.Replace(versionGo, "$TAG$", tag, 1)
 
+	if *dryRun {
+		fmt.Println(versionGo)
+		return
+	}
+
 	f, err := os.Create(fileName)
-	checkIfError(err)
+	checkIfError("creating version.go", err)
 	defer f.Close()
 
 	_, err = f.WriteString(versionGo)
-	checkIfError(err)
+	checkIfError("writing version.go", err)
 }
 
 func getRepository(repo string) *git.Repository {
@@ -208,12 +243,12 @@ func setTag(r *git.Repository, tag string) (bool, error) {
 	return true, nil
 }
 
-func checkIfError(err error) {
+func checkIfError(msg string, err error) {
 	if err == nil {
 		return
 	}
 
-	panic(err)
+	panic(fmt.Errorf("%v: %v", msg, err))
 }
 
 const versionTemplateStr = `/*
